requirements.txt
Django
psycopg2
djangorestframework
django-filter
dicttoxml
django-reversion